convert StateView to attached properties, update controller, rename to StateContainer
